requirements.txt
addict
diffusers
gradio
huggingface_hub
matplotlib
numpy
onnxruntime
Pillow
pycocotools
PyYAML
requests
setuptools
supervision
termcolor
timm
torch
torchvision
transformers
yapf
nltk
fairscale
litellm
openai
tqdm
pycocotools
opencv-python
matplotlib
pillow
scikit-learn
onnxruntime
onnx
ipykernel